requirements.txt
absl-py==1.4.0
appdirs==1.4.4
array-record==0.5.0
asttokens==2.4.1
astunparse==1.6.3
backports.functools-lru-cache==1.6.5
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
chex==0.1.85
click==8.1.3
comm==0.1.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dm-tree==0.1.8
docker-pycreds==0.4.0
etils==1.5.2
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
flax==0.7.5
fonttools==4.39.4
fsspec==2023.10.0
gast==0.4.0
gitdb==4.0.10
gitpython==3.1.31
google-auth-oauthlib==1.0.0
google-auth==2.19.1
google-pasta==0.2.0
googleapis-common-protos==1.61.0
grpcio==1.54.2
h5py==3.8.0
huggingface-hub==0.19.4
idna==3.4
imageio==2.33.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.17.2
jax==0.4.20
jaxlib==0.4.20
jedi==0.19.1
joblib==1.2.0
jupyter-client==8.5.0
jupyter-core==5.5.0
keras==2.15.0
kiwisolver==1.4.4
lazy-loader==0.3
libclang==16.0.0
markdown-it-py==3.0.0
markdown==3.4.3
markupsafe==2.1.3
matplotlib-inline==0.1.6
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu11==8.9.2.26
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu12==2.16.5
nvidia-nvjitlink-cu12==12.2.140
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.4
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.5.0
pip==24.0
platformdirs==3.11.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1-modules==0.3.0
pyasn1==0.5.0
pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0
pyzmq==25.1.0
regex==2023.10.3
requests-oauthlib==1.3.1
requests==2.31.0
rich==13.7.0
rsa==4.9
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.25.1
setproctitle==1.3.2
setuptools==67.8.0
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
stack-data==0.6.2
tensorboard-data-server==0.7.0
tensorboard==2.15.1
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.14.0
tensorflow-text==2.15.0
tensorflow==2.15.0
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
tensorrt==8.6.1.post1
tensorstore==0.1.50
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.12.9
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.2
typeguard==2.13.3
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wandb==0.16.3
wcwidth==0.2.9
werkzeug==2.3.6
wheel==0.38.4
wrapt==1.14.1
zipp==3.17.0